+++ examples/combo_box_table/combo_box/index.jsx
@@ -2,9 +2,7 @@ import { useState } from 'react';
 import {
   ComboBoxTable,
   useTokenSearch,
-  tokenHighlighter,
-  passThroughHighlighter,
-  highlightValue,
+  TokenHighlight,
   layoutMaxWidth,
   layoutMaxHeight,
   layoutColumnsAlignLeft,
@@ -17,12 +15,13 @@ function mapOption({ breed }) {
   return breed;
 }
 
-function highlighter(term, query, options, state) {
-  const { column: { name } } = state;
+function renderValue({ children }, { search, column: { name } }) {
   if (name === 'breed') {
-    return tokenHighlighter(term, query);
+    return (
+      <TokenHighlight search={search} label={children} />
+    );
   }
-  return passThroughHighlighter(term);
+  return children;
 }
 
 const onLayoutListBox = [layoutMaxWidth, layoutMaxHeight, layoutColumnsAlignLeft];
@@ -49,7 +48,7 @@ export function Example() {
         onLayoutListBox={onLayoutListBox}
         options={filteredOptions}
         columns={columns}
-        renderColumnValue={highlightValue(highlighter)}
+        renderColumnValue={renderValue}
         mapOption={mapOption}
       />