+++ examples/combo_box_table/combo_box_grouped/index.jsx
@@ -2,9 +2,7 @@ import { useState } from 'react';
 import {
   ComboBoxTable,
   useTokenSearch,
-  tokenHighlighter,
-  passThroughHighlighter,
-  highlightValue,
+  TokenHighlight,
   layoutMaxWidth,
   layoutMaxHeight,
   layoutColumnsAlignLeft,
@@ -21,12 +19,13 @@ function index({ breed }) {
   return breed;
 }
 
-function highlighter(term, query, options, state) {
-  const { column: { name } } = state;
+function renderValue({ children }, { search, column: { name } }) {
   if (name === 'breed') {
-    return tokenHighlighter(term, query);
+    return (
+      <TokenHighlight search={search} label={children} />
+    );
   }
-  return passThroughHighlighter(term);
+  return children;
 }
 
 const onLayoutListBox = [layoutMaxWidth, layoutMaxHeight, layoutColumnsAlignLeft];
@@ -52,7 +51,7 @@ export function Example() {
         onSearch={setSearch}
         options={filteredOptions}
         columns={columns}
-        renderColumnValue={highlightValue(highlighter)}
+        renderColumnValue={renderValue}
         mapOption={mapOption}
         onLayoutListBox={onLayoutListBox}
       />