+++ src/components/highlighters/delimited_highlight.jsx
@@ -1,6 +1,23 @@
-import { highlightValue } from './highlight_value';
+import PropTypes from 'prop-types';
+import { Highlight } from '../highlight';
 import { delimitedHighlighter } from '../../highlighters/delimited_highlighter';
 
-export function delimitedHighlight(options) {
-  return highlightValue(delimitedHighlighter, options);
+export function DelimitedHighlight({ label, inverse, start, end }) {
+  return (
+    <Highlight inverse={inverse}>
+      {delimitedHighlighter(label || '', { start, end })}
+    </Highlight>
+  );
 }
+
+DelimitedHighlight.propTypes = {
+  label: PropTypes.string,
+  start: PropTypes.string.isRequired,
+  end: PropTypes.string.isRequired,
+  inverse: PropTypes.bool,
+};
+
+DelimitedHighlight.defaultProps = {
+  label: '',
+  inverse: false,
+};