+++ src/components/highlighters/delimited_highlight.test.jsx
@@ -1,38 +1,60 @@
+/* eslint-disable quotes */
+
 import { render } from '@testing-library/react';
-import { delimitedHighlight } from './delimited_highlight';
+import { DelimitedHighlight } from './delimited_highlight';
 
-it('does not highlight with no children', () => {
-  const { container } = render((
-    delimitedHighlight({ start: '<em>', end: '</em>' })({ children: null }, { search: null }, {})
-  ));
-
-  expect(container).toContainHTML('<div></div>');
+it('highlights empty string', () => {
+  const { container } = render(<DelimitedHighlight label="" start="<em>" end="</em>" />);
+  expect(container).toMatchInlineSnapshot(`<div />`);
 });
 
-it('does not highlight with no search children', () => {
-  const { container } = render((
-    delimitedHighlight({ start: '<em>', end: '</em>' })({ children: 'foo' }, { search: null }, {})
-  ));
-
-  expect(container).toContainHTML('<div>foo</div>');
+it('highlights with no match', () => {
+  const { container } = render(<DelimitedHighlight label="foo" start="<em>" end="</em>" />);
+  expect(container).toMatchInlineSnapshot(`
+<div>
+  foo
+</div>
+`);
 });
 
-it('converts <em> in html strings to a highlight', () => {
-  const { container } = render((
-    delimitedHighlight({ start: '<em>', end: '</em>' })({ children: 'foo <em>bar</em> <em>foe</em>' }, { search: null }, { visuallyHiddenClassName: 'sr-only' })
-  ));
-
-  expect(container).toMatchSnapshot();
+it('highlights with a match', () => {
+  const { container } = render(<DelimitedHighlight label="foo <em>bar</em>" start="<em>" end="</em>" />);
+  expect(container).toMatchInlineSnapshot(`
+<div>
+  <span
+    class="visually-hidden visuallyhidden sr-only react-combo-boxes-sr-only"
+  >
+    foo bar
+  </span>
+  <span
+    aria-hidden="true"
+  >
+    foo 
+    <mark>
+      bar
+    </mark>
+  </span>
+</div>
+`);
 });
 
-it('inverses the highlight', () => {
-  const { container } = render((
-    delimitedHighlight({ start: '<em>', end: '</em>', inverse: true })(
-      { children: 'foo <em>bar</em> <em>foe</em>' },
-      { search: null },
-      { visuallyHiddenClassName: 'sr-only' },
-    )
-  ));
-
-  expect(container).toMatchSnapshot();
+it('inverses a highlight', () => {
+  const { container } = render(<DelimitedHighlight label="foo <em>bar</em>" start="<em>" end="</em>" inverse />);
+  expect(container).toMatchInlineSnapshot(`
+<div>
+  <span
+    class="visually-hidden visuallyhidden sr-only react-combo-boxes-sr-only"
+  >
+    foo bar
+  </span>
+  <span
+    aria-hidden="true"
+  >
+    <mark>
+      foo
+    </mark>
+     bar
+  </span>
+</div>
+`);
 });