+++ src/components/highlighters/index.js
@@ -1,6 +1,4 @@
-export { delimitedHighlight } from './delimited_highlight';
-export { passThroughHighlight } from './pass_through_highlight';
-export { prefixHighlight } from './prefix_highlight';
-export { substringHighlight } from './substring_highlight';
-export { tokenHighlight } from './token_highlight';
-export { highlightValue } from './highlight_value';
+export { DelimitedHighlight } from './delimited_highlight';
+export { PrefixHighlight } from './prefix_highlight';
+export { SubstringHighlight } from './substring_highlight';
+export { TokenHighlight } from './token_highlight';