+++ src/components/highlighters/prefix_highlight.jsx
@@ -1,6 +1,23 @@
-import { highlightValue } from './highlight_value';
+import PropTypes from 'prop-types';
+import { Highlight } from '../highlight';
 import { prefixHighlighter } from '../../highlighters/prefix_highlighter';
 
-export function prefixHighlight(options) {
-  return highlightValue(prefixHighlighter, options);
+export function PrefixHighlight({ label, search, inverse }) {
+  return (
+    <Highlight inverse={inverse}>
+      {prefixHighlighter(label || '', search || '')}
+    </Highlight>
+  );
 }
+
+PrefixHighlight.propTypes = {
+  label: PropTypes.string,
+  search: PropTypes.string,
+  inverse: PropTypes.bool,
+};
+
+PrefixHighlight.defaultProps = {
+  label: '',
+  search: '',
+  inverse: false,
+};